Renamed Tracker to Workload Generator (#185) (#205)
* Renamed Tracker to Workload_Processor


* Update import in test_index.py to adhere to pylint standards


* Rename workload_processor to workload_generator to be more apt
