+++ tests/workload_generator/index_test.py
@@ -23,7 +23,10 @@
 # under the License.
 
 from unittest import mock
-from osbenchmark.tracker.index import filter_ephemeral_index_settings, extract_index_mapping_and_settings, update_index_setting_parameters
+from osbenchmark.workload_generator.index import (
+    filter_ephemeral_index_settings,
+    extract_index_mapping_and_settings,
+    update_index_setting_parameters)
 
 
 def test_index_setting_filter():
